+++ lang/attribute/account/src/lib.rs
@@ -103,9 +103,9 @@ pub fn account(
         let discriminator_preimage = {
             // For now, zero copy accounts can't be namespaced.
             if namespace.is_empty() {
-                format!("account:{}", account_name)
+                format!("account:{account_name}")
             } else {
-                format!("{}:{}", namespace, account_name)
+                format!("{namespace}:{account_name}")
             }
         };
 
@@ -113,7 +113,7 @@ pub fn account(
         discriminator.copy_from_slice(
             &anchor_syn::hash::hash(discriminator_preimage.as_bytes()).to_bytes()[..8],
         );
-        format!("{:?}", discriminator).parse().unwrap()
+        format!("{discriminator:?}").parse().unwrap()
     };
 
     let owner_impl = {
@@ -280,9 +280,9 @@ pub fn derive_zero_copy_accessor(item: proc_macro::TokenStream) -> proc_macro::T
                     let field_name = field.ident.as_ref().unwrap();
 
                     let get_field: proc_macro2::TokenStream =
-                        format!("get_{}", field_name).parse().unwrap();
+                        format!("get_{field_name}").parse().unwrap();
                     let set_field: proc_macro2::TokenStream =
-                        format!("set_{}", field_name).parse().unwrap();
+                        format!("set_{field_name}").parse().unwrap();
 
                     quote! {
                         pub fn #get_field(&self) -> #accessor_ty {