+++ lang/syn/src/idl/relations.rs
@@ -11,7 +11,7 @@ pub fn parse(acc: &Field, seeds_feature: bool) -> Vec<String> {
         .flat_map(|s| match &s.join_target {
             Expr::Path(path) => path.path.segments.first().map(|l| l.ident.to_string()),
             _ => {
-                println!("WARNING: unexpected seed: {:?}", s);
+                println!("WARNING: unexpected seed: {s:?}");
                 None
             }
         })